running shell operation in the dbt transform flow and removing the exception thrown in flows

+++ proxy/flows.py
@@ -5,12 +5,20 @@ from fastapi import HTTPException
 from prefect import flow
 from prefect_airbyte.flows import run_connection_sync
 from prefect_airbyte import AirbyteConnection
-from prefect_dbt.cli.commands import DbtCoreOperation
+from prefect_dbt.cli.commands import DbtCoreOperation, ShellOperation
 from logger import logger
+
+
+# django prefect block names
+AIRBYTESERVER = "Airbyte Server"
+AIRBYTECONNECTION = "Airbyte Connection"
+SHELLOPERATION = "Shell Operation"
+DBTCORE = "dbt Core Operation"
 
 
 @flow
 def run_airbyte_connection_flow(block_name: str):
+    # pylint: disable=broad-exception-caught
     """Prefect flow to run airbyte connection"""
     try:
         airbyte_connection = AirbyteConnection.load(block_name)
@@ -18,14 +26,15 @@ def run_airbyte_connection_flow(block_name: str):
         logger.info("airbyte connection sync result=")
         logger.info(result)
         return result
-    except Exception as error:
+    except Exception as error:  # pylint: disable=broad-exception-caught
         # logger.exception(error)
         logger.error(str(error))  # "Job <num> failed."
-        raise HTTPException(status_code=400, detail=str(error)) from error
+        # raise HTTPException(status_code=400, detail=str(error)) from error
 
 
 @flow
 def run_dbtcore_flow(block_name: str):
+    # pylint: disable=broad-exception-caught
     """Prefect flow to run dbt"""
     try:
         dbt_op = DbtCoreOperation.load(block_name)
@@ -33,13 +42,13 @@ def run_dbtcore_flow(block_name: str):
             os.unlink(dbt_op.profiles_dir / "profiles.yml")
         return dbt_op.run()
     except Exception as error:
-        # logger.exception(error)
-        logger.error("FAILED")
-        raise HTTPException(status_code=400, detail=str(error)) from error
+        logger.exception(error)
+        # raise HTTPException(status_code=400, detail=str(error)) from error
 
 
 @flow
 def deployment_schedule_flow(airbyte_blocks: list, dbt_blocks: list):
+    # pylint: disable=broad-exception-caught
     """A general flow function that will help us create deployments"""
     # sort the dbt blocks by seq
     dbt_blocks.sort(key=lambda blk: blk["seq"])
@@ -54,15 +63,21 @@ def deployment_schedule_flow(airbyte_blocks: list, dbt_blocks: list):
             run_connection_sync(airbyte_connection)
         except Exception as error:
             logger.exception(error)
-            raise HTTPException(status_code=400, detail=str(error)) from error
+            # raise HTTPException(status_code=400, detail=str(error)) from error
 
     # run dbt blocks
     for block in dbt_blocks:
-        dbt_op = DbtCoreOperation.load(block["blockName"])
-        if os.path.exists(dbt_op.profiles_dir / "profiles.yml"):
-            os.unlink(dbt_op.profiles_dir / "profiles.yml")
-        try:
-            dbt_op.run()
-        except Exception as error:
-            logger.exception(error)
-            raise HTTPException(status_code=400, detail=str(error)) from error
+        if block["blockType"] == SHELLOPERATION:
+            shell_op = ShellOperation.load(block["blockName"])
+            shell_op.run()
+            continue
+
+        if block["blockType"] == DBTCORE:
+            dbt_op = DbtCoreOperation.load(block["blockName"])
+            if os.path.exists(dbt_op.profiles_dir / "profiles.yml"):
+                os.unlink(dbt_op.profiles_dir / "profiles.yml")
+            try:
+                dbt_op.run()
+            except Exception as error:
+                logger.exception(error)
+                # raise HTTPException(status_code=400, detail=str(error)) from error
